refactor(git): extract commit-drafting commands into git_commit.py
