+++ .rrt/tree.lock.toml
@@ -1,10 +1,10 @@
 [meta]
-generated_at = "2026-07-05T07:42:48+00:00"
+generated_at = "2026-07-05T07:51:17+00:00"
 rrt_version = "1.10.1"
 
 [snapshot]
-entry_count = 323
-tree_hash = "sha256:ed568d5f915f8c6c4f105b460d5711a9afb8d7ec357a59f309e34776aa10252d"
-updated_at = "2026-07-05T07:42:48+00:00"
+entry_count = 325
+tree_hash = "sha256:e2f6104f71075a224f18e30853ed99e9b45ff703a3d08e475f653f6835a0cb9c"
+updated_at = "2026-07-05T07:51:17+00:00"
 ignored_count = 32
 phantom_empty_dirs = []
